backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.0.1
email-validator==2.1.0

# Logging
python-json-logger==2.0.7

# XML Processing
zeep==4.2.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
